+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/graph/DependencyGraphNode.kt
@@ -42,11 +42,11 @@ internal data class DependencyGraphNode(
   val graphExtensions: Map<IrTypeKey, List<GraphExtensionAccessor>>,
   val scopes: Set<IrAnnotation>,
   val aggregationScopes: Set<ClassId>,
-  val providerFactories: List<Pair<IrTypeKey, ProviderFactory>>,
+  val providerFactories: Map<IrTypeKey, ProviderFactory>,
   // Types accessible via this graph (includes inherited)
   // Dagger calls these "provision methods", but that's a bit vague IMO
   val accessors: List<GraphAccessor>,
-  val bindsCallables: Set<BindsCallable>,
+  val bindsCallables: Map<IrTypeKey, BindsCallable>,
   val multibindsCallables: Set<MultibindsCallable>,
   val optionalKeys: Map<IrTypeKey, Set<BindsOptionalOfCallable>>,
   /** Binding containers that need a managed instance. */