+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/BindingContainerTransformer.kt
@@ -49,6 +49,7 @@ import dev.zacsweers.metro.compiler.ir.subcomponentsArgument
 import dev.zacsweers.metro.compiler.ir.thisReceiverOrFail
 import dev.zacsweers.metro.compiler.ir.toClassReferences
 import dev.zacsweers.metro.compiler.ir.toProto
+import dev.zacsweers.metro.compiler.ir.transformMultiboundQualifier
 import dev.zacsweers.metro.compiler.ir.writeDiagnostic
 import dev.zacsweers.metro.compiler.isWordPrefixRegex
 import dev.zacsweers.metro.compiler.mapNotNullToSet
@@ -822,10 +823,12 @@ internal class BindingContainerTransformer(context: IrMetroContext) : IrMetroCon
                   )
                   return null
                 }
+                val transformedTypeKey = typeKey.transformMultiboundQualifier(annotations)
                 providerFactories[callableId] =
                   ProviderFactory.Dagger(
                     factoryClass = factoryClass.owner,
-                    typeKey = typeKey,
+                    typeKey = transformedTypeKey,
+                    rawTypeKey = typeKey,
                     callableId = callableId,
                     annotations = annotations,
                     parameters = parameters,